+++ CHANGELOG.rst
@@ -6,7 +6,8 @@ Changelog
 Next
 ----
 
-- Breaking change: ``sphinx-prompt`` is now an optional dependency.
+- Breaking change: The required Sphinx version is at least 3.0.
+- ``sphinx-prompt`` is now an optional dependency.
   Thanks go to @dgarcia360 for this change.
 
 2020.09.30.0