+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 Next
 ----
 
-- Breaking change: The required Sphinx version is at least 3.0.
+- Breaking change: The required Sphinx version is at least 4.0.
 - ``sphinx-prompt`` is now an optional dependency.
   Thanks go to @dgarcia360 for this change.